requirements.txt
albumentations
einops==0.8.0
h5py==3.9.0
joblib==1.4.2
kornia==0.7.2
loguru==0.5.3
matplotlib==3.7.5
numpy==1.24.3
opencv_python==4.9.0.80
Pillow
Pillow
pytorch_lightning==1.3.5
setuptools==68.2.2
timm==0.9.16
tqdm==4.66.2
yacs==0.1.8
